+++ src/elastic-san/azext_elastic_san/aaz/latest/elastic_san/volume/_restore.py
@@ -12,27 +12,28 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "elastic-san volume snapshot list",
+    "elastic-san volume restore",
+    is_preview=True,
 )
-class List(AAZCommand):
-    """List Snapshots in a VolumeGroup or List Snapshots by Volume (name) in a VolumeGroup using filter
+class Restore(AAZCommand):
+    """Restore Soft Deleted Volumes.
 
-    :example: snapshot list
-        az elastic-san volume snapshot list -g "rg" -e "san_name" -v "vg_name"
+    :example: Restore a deleted volume
+        az elastic-san volume restore -g rg_name -e san_name -v volume_group_name -n deleted_volume_name
     """
 
     _aaz_info = {
         "version": "2024-07-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.elasticsan/elasticsans/{}/volumegroups/{}/snapshots", "2024-07-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.elasticsan/elasticsans/{}/volumegroups/{}/volumes/{}/restore", "2024-07-01-preview"],
         ]
     }
 
-    AZ_SUPPORT_PAGINATION = True
+    AZ_SUPPORT_NO_WAIT = True
 
     def _handler(self, command_args):
         super()._handler(command_args)
-        return self.build_paging(self._execute_operations, self._output)
+        return self.build_lro_poller(self._execute_operations, self._output)
 
     _args_schema = None
 
@@ -49,6 +50,7 @@ class List(AAZCommand):
             options=["-e", "--elastic-san", "--elastic-san-name"],
             help="The name of the ElasticSan.",
             required=True,
+            id_part="name",
             fmt=AAZStrArgFormat(
                 pattern="^[A-Za-z0-9]+((-|_)[a-z0-9A-Z]+)*$",
                 max_length=24,
@@ -62,21 +64,29 @@ class List(AAZCommand):
             options=["-v", "--volume-group", "--volume-group-name"],
             help="The name of the VolumeGroup.",
             required=True,
+            id_part="child_name_1",
             fmt=AAZStrArgFormat(
                 pattern="^[A-Za-z0-9]+((-|_)[a-z0-9A-Z]+)*$",
                 max_length=63,
                 min_length=3,
             ),
         )
-        _args_schema.filter = AAZStrArg(
-            options=["--filter"],
-            help="Specify `$filter='volumeName eq <volume name>'` to filter on volume.",
+        _args_schema.volume_name = AAZStrArg(
+            options=["-n", "--name", "--volume-name"],
+            help="The name of the Volume.",
+            required=True,
+            id_part="child_name_2",
+            fmt=AAZStrArgFormat(
+                pattern="^[a-z0-9]+(-[a-z0-9A-Z]+)*$",
+                max_length=63,
+                min_length=3,
+            ),
         )
         return cls._args_schema
 
     def _execute_operations(self):
         self.pre_operations()
-        self.VolumeSnapshotsListByVolumeGroup(ctx=self.ctx)()
+        yield self.RestoreVolume(ctx=self.ctx)()
         self.post_operations()
 
     @register_callback
@@ -88,31 +98,46 @@ class List(AAZCommand):
         pass
 
     def _output(self, *args, **kwargs):
-        result = self.deserialize_output(self.ctx.vars.instance.value, client_flatten=True)
-        next_link = self.deserialize_output(self.ctx.vars.instance.next_link)
-        return result, next_link
+        result = self.deserialize_output(self.ctx.vars.instance, client_flatten=True)
+        return result
 
-    class VolumeSnapshotsListByVolumeGroup(AAZHttpOperation):
+    class RestoreVolume(AAZHttpOperation):
         CLIENT_TYPE = "MgmtClient"
 
         def __call__(self, *args, **kwargs):
             request = self.make_request()
             session = self.client.send_request(request=request, stream=False, **kwargs)
+            if session.http_response.status_code in [202]:
+                return self.client.build_lro_polling(
+                    self.ctx.args.no_wait,
+                    session,
+                    self.on_200,
+                    self.on_error,
+                    lro_options={"final-state-via": "location"},
+                    path_format_arguments=self.url_parameters,
+                )
             if session.http_response.status_code in [200]:
-                return self.on_200(session)
+                return self.client.build_lro_polling(
+                    self.ctx.args.no_wait,
+                    session,
+                    self.on_200,
+                    self.on_error,
+                    lro_options={"final-state-via": "location"},
+                    path_format_arguments=self.url_parameters,
+                )
 
             return self.on_error(session.http_response)
 
         @property
         def url(self):
             return self.client.format_url(
-                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ElasticSan/elasticSans/{elasticSanName}/volumegroups/{volumeGroupName}/snapshots",
+                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ElasticSan/elasticSans/{elasticSanName}/volumegroups/{volumeGroupName}/volumes/{volumeName}/restore",
                 **self.url_parameters
             )
 
         @property
         def method(self):
-            return "GET"
+            return "POST"
 
         @property
         def error_format(self):
@@ -137,15 +162,16 @@ class List(AAZCommand):
                     "volumeGroupName", self.ctx.args.volume_group_name,
                     required=True,
                 ),
+                **self.serialize_url_param(
+                    "volumeName", self.ctx.args.volume_name,
+                    required=True,
+                ),
             }
             return parameters
 
         @property
         def query_parameters(self):
             parameters = {
-                **self.serialize_query_param(
-                    "$filter", self.ctx.args.filter,
-                ),
                 **self.serialize_query_param(
                     "api-version", "2024-07-01-preview",
                     required=True,
@@ -180,58 +206,80 @@ class List(AAZCommand):
             cls._schema_on_200 = AAZObjectType()
 
             _schema_on_200 = cls._schema_on_200
-            _schema_on_200.next_link = AAZStrType(
-                serialized_name="nextLink",
+            _schema_on_200.id = AAZStrType(
                 flags={"read_only": True},
             )
-            _schema_on_200.value = AAZListType()
-
-            value = cls._schema_on_200.value
-            value.Element = AAZObjectType()
-
-            _element = cls._schema_on_200.value.Element
-            _element.id = AAZStrType(
+            _schema_on_200.name = AAZStrType(
                 flags={"read_only": True},
             )
-            _element.name = AAZStrType(
-                flags={"read_only": True},
-            )
-            _element.properties = AAZObjectType(
+            _schema_on_200.properties = AAZObjectType(
                 flags={"required": True, "client_flatten": True},
             )
-            _element.system_data = AAZObjectType(
+            _schema_on_200.system_data = AAZObjectType(
                 serialized_name="systemData",
                 flags={"read_only": True},
             )
-            _element.type = AAZStrType(
+            _schema_on_200.type = AAZStrType(
                 flags={"read_only": True},
             )
 
-            properties = cls._schema_on_200.value.Element.properties
+            properties = cls._schema_on_200.properties
             properties.creation_data = AAZObjectType(
                 serialized_name="creationData",
-                flags={"required": True},
+            )
+            properties.managed_by = AAZObjectType(
+                serialized_name="managedBy",
             )
             properties.provisioning_state = AAZStrType(
                 serialized_name="provisioningState",
                 flags={"read_only": True},
             )
-            properties.source_volume_size_gi_b = AAZIntType(
-                serialized_name="sourceVolumeSizeGiB",
-                flags={"read_only": True},
-            )
-            properties.volume_name = AAZStrType(
-                serialized_name="volumeName",
-                flags={"read_only": True},
-            )
-
-            creation_data = cls._schema_on_200.value.Element.properties.creation_data
-            creation_data.source_id = AAZStrType(
-                serialized_name="sourceId",
+            properties.size_gi_b = AAZIntType(
+                serialized_name="sizeGiB",
                 flags={"required": True},
             )
+            properties.storage_target = AAZObjectType(
+                serialized_name="storageTarget",
+                flags={"read_only": True},
+            )
+            properties.volume_id = AAZStrType(
+                serialized_name="volumeId",
+                flags={"read_only": True},
+            )
 
-            system_data = cls._schema_on_200.value.Element.system_data
+            creation_data = cls._schema_on_200.properties.creation_data
+            creation_data.create_source = AAZStrType(
+                serialized_name="createSource",
+            )
+            creation_data.source_id = AAZStrType(
+                serialized_name="sourceId",
+            )
+
+            managed_by = cls._schema_on_200.properties.managed_by
+            managed_by.resource_id = AAZStrType(
+                serialized_name="resourceId",
+            )
+
+            storage_target = cls._schema_on_200.properties.storage_target
+            storage_target.provisioning_state = AAZStrType(
+                serialized_name="provisioningState",
+                flags={"read_only": True},
+            )
+            storage_target.status = AAZStrType()
+            storage_target.target_iqn = AAZStrType(
+                serialized_name="targetIqn",
+                flags={"read_only": True},
+            )
+            storage_target.target_portal_hostname = AAZStrType(
+                serialized_name="targetPortalHostname",
+                flags={"read_only": True},
+            )
+            storage_target.target_portal_port = AAZIntType(
+                serialized_name="targetPortalPort",
+                flags={"read_only": True},
+            )
+
+            system_data = cls._schema_on_200.system_data
             system_data.created_at = AAZStrType(
                 serialized_name="createdAt",
             )
@@ -254,8 +302,8 @@ class List(AAZCommand):
             return cls._schema_on_200
 
 
-class _ListHelper:
-    """Helper class for List"""
+class _RestoreHelper:
+    """Helper class for Restore"""
 
 
-__all__ = ["List"]
+__all__ = ["Restore"]